Separate Settings implementation for <NC20 and >NC20
To remove and replace as soon as NC19 is out of support on Forms.
